+++ ext/duckdb/scalar_function_bind_info.c
@@ -7,6 +7,7 @@ static VALUE allocate(VALUE klass);
 static size_t memsize(const void *p);
 static VALUE rbduckdb_scalar_function_bind_info_argument_count(VALUE self);
 static VALUE rbduckdb_scalar_function_bind_info_set_error(VALUE self, VALUE error);
+static VALUE rbduckdb_scalar_function_bind_info_get_argument(VALUE self, VALUE index);
 
 static const rb_data_type_t scalar_function_bind_info_data_type = {
     "DuckDB/ScalarFunction/BindInfo",
@@ -66,6 +67,20 @@ static VALUE rbduckdb_scalar_function_bind_info_set_error(VALUE self, VALUE erro
     return self;
 }
 
+/*
+ * call-seq:
+ *   bind_info._get_argument(index) -> DuckDB::Expression
+ *
+ * Returns the expression at the given argument index.
+ * Called internally by +get_argument+ after index validation.
+ */
+static VALUE rbduckdb_scalar_function_bind_info_get_argument(VALUE self, VALUE index) {
+    rubyDuckDBScalarFunctionBindInfo *ctx;
+    TypedData_Get_Struct(self, rubyDuckDBScalarFunctionBindInfo, &scalar_function_bind_info_data_type, ctx);
+    duckdb_expression expr = duckdb_scalar_function_bind_get_argument(ctx->bind_info, (idx_t)NUM2ULL(index));
+    return rbduckdb_expression_new(expr);
+}
+
 void rbduckdb_init_duckdb_scalar_function_bind_info(void) {
 #if 0
     VALUE mDuckDB = rb_define_module("DuckDB");
@@ -74,4 +89,5 @@ void rbduckdb_init_duckdb_scalar_function_bind_info(void) {
     rb_define_alloc_func(cDuckDBScalarFunctionBindInfo, allocate);
     rb_define_method(cDuckDBScalarFunctionBindInfo, "argument_count", rbduckdb_scalar_function_bind_info_argument_count, 0);
     rb_define_method(cDuckDBScalarFunctionBindInfo, "set_error", rbduckdb_scalar_function_bind_info_set_error, 1);
+    rb_define_private_method(cDuckDBScalarFunctionBindInfo, "_get_argument", rbduckdb_scalar_function_bind_info_get_argument, 1);
 }